+++ pontos/version/__init__.py
@@ -28,12 +28,6 @@ from .version import (
     get_version_from_pyproject_toml,
 )
 from .cmake_version import CMakeVersionParser, CMakeVersionCommand
-from .helper import (
-    calculate_calendar_version,
-    get_current_version,
-    get_next_patch_version,
-    get_next_dev_version,
-)
 
 
 def main(leave=True, args=None):
@@ -55,12 +49,8 @@ def main(leave=True, args=None):
 
 __all__ = [
     '__version__',
-    'calculate_calendar_version',
     'CMakeVersionCommand',
     'get_version_from_pyproject_toml',
-    'get_current_version',
-    'get_next_patch_version',
-    'get_next_dev_version',
     'is_version_pep440_compliant',
     'PontosVersionCommand',
     'safe_version',